address some PR comments
