+++ package.nls.json
@@ -100,7 +100,7 @@
 "mssql.format.placeCommasBeforeNextStatement":"should commas be placed at the beginning of each statement in a list e.g. ', mycolumn2' instead of at the end e.g. 'mycolumn1,'",
 "mssql.format.placeSelectStatementReferencesOnNewLine":"Should references to objects in a select statements be split into separate lines? E.g. for 'SELECT C1, C2 FROM T1' both C1 and C2 will be on separate lines",
 "mssql.applyLocalization":"[Optional] Configuration options for localizing into VSCode's configured locale (must restart VSCode for settings to take effect)",
-"mssql.resultsGrid.autoSizeColumns": "Auto size the columns width on initial results. Could have performance problems with a large number of columns or large cells",
+"mssql.resultsGrid.autoSizeColumns": "Automatically adjust the column widths based on the visible rows in the result set. Could have performance problems with a large number of columns or large cells",
 "mssql.query.displayBitAsNumber":"Should BIT columns be displayed as numbers (1 or 0)? If false, BIT columns will be displayed as 'true' or 'false'",
 "mssql.intelliSense.enableIntelliSense":"Should IntelliSense be enabled",
 "mssql.intelliSense.enableErrorChecking":"Should IntelliSense error checking be enabled",
