+++ src/reactviews/pages/QueryResult/table/utils.ts
@@ -75,18 +75,3 @@ export function tryCombineSelections(selections: ISlickRange[]): ISlickRange[] {
     }
     return [unifiedSelection];
 }
-
-export function deepClone<T>(obj: T): T {
-    if (!obj || typeof obj !== "object") {
-        return obj;
-    }
-    if (obj instanceof RegExp) {
-        return obj;
-    }
-    const result: any = Array.isArray(obj) ? [] : {};
-    Object.entries(obj).forEach(([key, value]) => {
-        result[key] =
-            value && typeof value === "object" ? deepClone(value) : value;
-    });
-    return result;
-}
